New subpackage for gopacket support
This package includes three files that allow easy decoding using the
google/gopacket package. In order to use this, you will need to register
the header you want to use (either Short or fragment) to the corresponding UDP
port (e.g. LCM multicast 7667) by using gopacket/layers'
RegisterUDPPortLayerType function. Further improvements could include a wrap of
this call in the code of this library.

It is not yet clear how this will behave exactly in regard to the
different decoding functionalities of the library, as it was mainly used
in combination with packet.Layer method.