Added lcm_struct->pkg for multi file lcmgen
The lcmgen->package only saved the last package, which is problematic
when lcmgen is started with multiple .lcm files. If each is within a
different package it should be possible to access the correct package
in which the struct was defined. Each struct now has its own package
member which contains this info.