(lcmheader.go) fragmented header computation fix
This patch fixes a logic error that caused unpopulated fields in the
header when decoding fragmented headers. This logic error tried to
extract the channel at a fixed offset of 8 bytes, which is only working
when used with an LCM short header. We are now extracting the channel
name at a relative offset + only for short headers and the very first
frame of a fragmented message, as described in the LCM spefication.